--- a/dask_sql/__init__.py
+++ b/dask_sql/__init__.py
@@ -1,6 +1,9 @@
 from ._version import get_version
 from .cmd import cmd_loop
 from .context import Context
+from .datacontainer import Statistics
 from .server.app import run_server
 
 __version__ = get_version()
+
+__all__ = [__version__, cmd_loop, Context, run_server, Statistics]