--- a/dask_sql/java.py
+++ b/dask_sql/java.py
@@ -85,18 +85,6 @@ com = jpype.JPackage("com")
 org = jpype.JPackage("org")
 java = jpype.JPackage("java")
 
-DaskTable = com.dask.sql.schema.DaskTable
-DaskAggregateFunction = com.dask.sql.schema.DaskAggregateFunction
-DaskScalarFunction = com.dask.sql.schema.DaskScalarFunction
-DaskSchema = com.dask.sql.schema.DaskSchema
-RelationalAlgebraGenerator = com.dask.sql.application.RelationalAlgebraGenerator
-RelationalAlgebraGeneratorBuilder = (
-    com.dask.sql.application.RelationalAlgebraGeneratorBuilder
-)
-SqlTypeName = org.apache.calcite.sql.type.SqlTypeName
-ValidationException = org.apache.calcite.tools.ValidationException
-SqlParseException = org.apache.calcite.sql.parser.SqlParseException
-
 
 def get_java_class(instance):
     """Get the stringified class name of a java object"""