--- a/dask_sql/physical/rel/base.py
+++ b/dask_sql/physical/rel/base.py
@@ -38,6 +38,8 @@ class BaseRelPlugin:
         """
         field_names = [str(x) for x in row_type.getFieldNames()]
 
+        logger.debug(f"Renaming {cc.columns} to {field_names}")
+
         cc = cc.rename(columns=dict(zip(cc.columns, field_names)))
 
         # TODO: We can also check for the types here and do any conversions if needed