--- a/dask_sql/physical/rel/logical/__init__.py
+++ b/dask_sql/physical/rel/logical/__init__.py
@@ -1,23 +1,25 @@
-from .aggregate import LogicalAggregatePlugin
-from .filter import LogicalFilterPlugin
-from .join import LogicalJoinPlugin
-from .project import LogicalProjectPlugin
+from .aggregate import DaskAggregatePlugin
+from .filter import DaskFilterPlugin
+from .join import DaskJoinPlugin
+from .limit import DaskLimitPlugin
+from .project import DaskProjectPlugin
 from .sample import SamplePlugin
-from .sort import LogicalSortPlugin
-from .table_scan import LogicalTableScanPlugin
-from .union import LogicalUnionPlugin
-from .values import LogicalValuesPlugin
-from .window import LogicalWindowPlugin
+from .sort import DaskSortPlugin
+from .table_scan import DaskTableScanPlugin
+from .union import DaskUnionPlugin
+from .values import DaskValuesPlugin
+from .window import DaskWindowPlugin
 
 __all__ = [
-    LogicalAggregatePlugin,
-    LogicalFilterPlugin,
-    LogicalJoinPlugin,
-    LogicalProjectPlugin,
-    LogicalSortPlugin,
-    LogicalTableScanPlugin,
-    LogicalUnionPlugin,
-    LogicalValuesPlugin,
-    LogicalWindowPlugin,
+    DaskAggregatePlugin,
+    DaskFilterPlugin,
+    DaskJoinPlugin,
+    DaskLimitPlugin,
+    DaskProjectPlugin,
+    DaskSortPlugin,
+    DaskTableScanPlugin,
+    DaskUnionPlugin,
+    DaskValuesPlugin,
+    DaskWindowPlugin,
     SamplePlugin,
 ]