--- a/dask_sql/physical/rel/logical/project.py
+++ b/dask_sql/physical/rel/logical/project.py
@@ -11,14 +11,14 @@ from dask_sql.utils import new_temporary_column
 logger = logging.getLogger(__name__)
 
 
-class LogicalProjectPlugin(BaseRelPlugin):
+class DaskProjectPlugin(BaseRelPlugin):
     """
-    A LogicalProject is used to
+    A DaskProject is used to
     (a) apply expressions to the columns and
     (b) only select a subset of the columns
     """
 
-    class_name = "org.apache.calcite.rel.logical.LogicalProject"
+    class_name = "com.dask.sql.nodes.DaskProject"
 
     def convert(
         self, rel: "org.apache.calcite.rel.RelNode", context: "dask_sql.Context"