--- a/dask_sql/physical/rel/logical/sort.py
+++ b/dask_sql/physical/rel/logical/sort.py
@@ -24,14 +24,13 @@ class DaskSortPlugin(BaseRelPlugin):
             for x in sort_collation
         ]
 
-        if sort_columns:
-            ASCENDING = org.apache.calcite.rel.RelFieldCollation.Direction.ASCENDING
-            FIRST = org.apache.calcite.rel.RelFieldCollation.NullDirection.FIRST
-            sort_ascending = [x.getDirection() == ASCENDING for x in sort_collation]
-            sort_null_first = [x.nullDirection == FIRST for x in sort_collation]
+        ASCENDING = org.apache.calcite.rel.RelFieldCollation.Direction.ASCENDING
+        FIRST = org.apache.calcite.rel.RelFieldCollation.NullDirection.FIRST
+        sort_ascending = [x.getDirection() == ASCENDING for x in sort_collation]
+        sort_null_first = [x.nullDirection == FIRST for x in sort_collation]
 
-            df = df.persist()
-            df = apply_sort(df, sort_columns, sort_ascending, sort_null_first)
+        df = df.persist()
+        df = apply_sort(df, sort_columns, sort_ascending, sort_null_first)
 
         cc = self.fix_column_to_row_type(cc, rel.getRowType())
         # No column type has changed, so no need to cast again