--- a/dask_sql/physical/rel/logical/union.py
+++ b/dask_sql/physical/rel/logical/union.py
@@ -5,13 +5,13 @@ from dask_sql.physical.rel.base import BaseRelPlugin
 from dask_sql.physical.rex import RexConverter
 
 
-class LogicalUnionPlugin(BaseRelPlugin):
+class DaskUnionPlugin(BaseRelPlugin):
     """
-    LogicalUnion is used on UNION clauses.
+    DaskUnion is used on UNION clauses.
     It just concatonates the two data frames.
     """
 
-    class_name = "org.apache.calcite.rel.logical.LogicalUnion"
+    class_name = "com.dask.sql.nodes.DaskUnion"
 
     def convert(
         self, rel: "org.apache.calcite.rel.RelNode", context: "dask_sql.Context"