Related to #183.
This PR introduces real cost-based optimization into dask-sql using the Calcite volcano planner.
For this, I did three steps:
* refactored the Relation`alAlgebraGenerator` class into smaller sub-classes to make the handing and go away from the "default" `Framework` program
* Put a volcano after the hep-planner, which is allowed to choose rules based on the costs
* give the user the possibility to add custom statistics to a table - currently, it is only the row count but we might want to add more later.

The `create_table` function has now an additional parameter `statistics`, where you can give a `dask_sql.Statistics` object. In my very first small tests, I could not find a relational algebra which is now optimized differently - but that does not mean there is none. Maybe we can find some benchmark use-cases.