@nils-braun 2021-08-30
Hi @rajagurunath! I know, this is quite a large PR (and I am sorry for that) and I do not expect you to read through all of it - but maybe you find the time to have a quick look? I am more or less confident that I did not screw it up completely (because I only changed minor things in the tests and they still work), but I would feel a lot better of you also have a rough look :-) Thanks!