@rajagurunath 2021-08-31
Hi @nils-braun,

Really Amazing work, kudos to you !!! 👏 

Not able to understand all the things, but able to correlate ~ 60% of things, And I have tested in my local,  as you mentioned it doesn't break anything and works like a charm. Dask Nodes and their respective Dask Rules are looking clean and elegant from Java side. 

I compared the plan of SQL, with and without cost-based optimization able to see effect of the row count in the plan, definitely it will be a great value add for the dask-sql users. 

**With Cost Based optimization plan** 

```
DaskProject(sepal_length=[$0], sepal_width=[$1], petal_length=[$2], petal_width=[$3], species=[$4]): rowcount = 22.5, cumulative cost = {256.25 rows, 314.5 cpu, 0.0 io}, id = 273
  DaskJoin(condition=[AND(=($4, $5), =($0, $6))], joinType=[inner]): rowcount = 22.5, cumulative cost = {233.75 rows, 202.0 cpu, 0.0 io}, id = 272
    DaskTableScan(table=[[root, iris]]): rowcount = 100.0, cumulative cost = {100.0 rows, 101.0 cpu, 0.0 io}, id = 249
    DaskAggregate(group=[{4}], sepal_length=[MAX($0)]): rowcount = 10.0, cumulative cost = {111.25 rows, 101.0 cpu, 0.0 io}, id = 271
      DaskTableScan(table=[[root, iris]]): rowcount = 100.0, cumulative cost = {100.0 rows, 101.0 cpu, 0.0 io}, id = 249
```
And time to complete the above join :)
`119 ms ± 2.14 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)`


**Without Cost Based optimization plan** 

```
LogicalProject(sepal_length=[$0], sepal_width=[$1], petal_length=[$2], petal_width=[$3], species=[$4])
  LogicalJoin(condition=[AND(=($4, $5), =($0, $6))], joinType=[inner])
    LogicalTableScan(table=[[root, iris]])
    LogicalAggregate(group=[{4}], sepal_length=[MAX($0)])
      LogicalTableScan(table=[[root, iris]])
```
`136 ms ± 12.9 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)`

Not able to do benchmark completely but with help of `dask-sql Feature-Overview.ipynb`, I am able to see some improvement and I am super exicited about that. (even the default row count was set as 100, but able to see some improvement in the above join query)

So As you have mentioned we will be adding more statistics (min, max, avg ) to the plan right? And how about the `ANALYZE` statement to add those statistics to each table. (maybe we will create a new sub-issue  to track this )